move type files
