+++ src/lib/app/eventTypes.gen.ts
