+++ src/lib/app/eventTypes.ts
@@ -1,4 +1,4 @@
-// generated by src/lib/ui/events.gen.ts
+// generated by src/lib/app/eventTypes.gen.ts
 
 import type {Readable} from 'svelte/store';
 
@@ -426,4 +426,4 @@ export interface UiHandlers {
 	select_space: (ctx: DispatchContext<SelectSpaceParamsType, void>) => void;
 }
 
-// generated by src/lib/ui/events.gen.ts
+// generated by src/lib/app/eventTypes.gen.ts
